+++ mne/utils.py
@@ -296,13 +296,28 @@ def check_random_state(seed):
                      ' instance' % seed)
 
 
-def split_list(l, n):
-    """Split list in n (approx) equal pieces."""
+def split_list(l, n, idx=False):
+    """Split list in n (approx) equal pieces, possibly giving indices."""
     n = int(n)
-    sz = len(l) // n
+    tot = len(l)
+    sz = tot // n
+    start = stop = 0
     for i in range(n - 1):
-        yield l[i * sz:(i + 1) * sz]
-    yield l[(n - 1) * sz:]
+        stop += sz
+        yield (np.arange(start, stop), l[start:stop]) if idx else l[start:stop]
+        start += sz
+    yield (np.arange(start, tot), l[start:]) if idx else l[start]
+
+
+def array_split_idx(ary, indices_or_sections, axis=0, n_per_split=1):
+    """Do what numpy.array_split does, but add indices."""
+    # this only works for indices_or_sections as int
+    indices_or_sections = _ensure_int(indices_or_sections)
+    ary_split = np.array_split(ary, indices_or_sections, axis=axis)
+    idx_split = np.array_split(np.arange(ary.shape[axis]), indices_or_sections)
+    idx_split = (np.arange(sp[0] * n_per_split, (sp[-1] + 1) * n_per_split)
+                 for sp in idx_split)
+    return zip(idx_split, ary_split)
 
 
 def create_chunks(sequence, size):
@@ -1709,7 +1724,7 @@ class ProgressBar(object):
     """
 
     spinner_symbols = ['|', '/', '-', '\\']
-    template = '\r[{0}{1}] {2:.02f}% {4} {3}   '
+    template = '\r[{0}{1}] {2:6.02f}% {4} {3}   '
 
     def __init__(self, max_value, initial_value=0, mesg='', max_chars='auto',
                  progress_character='.', spinner=False,
@@ -1719,13 +1734,15 @@ class ProgressBar(object):
             self.max_value = len(max_value)
             self.iterable = max_value
         else:
-            self.max_value = float(max_value)
+            self.max_value = max_value
             self.iterable = None
         self.mesg = mesg
         self.progress_character = progress_character
         self.spinner = spinner
         self.spinner_index = 0
         self.n_spinner = len(self.spinner_symbols)
+        if verbose_bool == 'auto':
+            verbose_bool = True if logger.level <= logging.INFO else False
         self._do_print = verbose_bool
         self.cur_time = time.time()
         if max_total_width == 'auto':
@@ -1735,6 +1752,10 @@ class ProgressBar(object):
             max_chars = min(max(max_total_width - 40, 10), 60)
         self.max_chars = int(max_chars)
         self.cur_rate = 0
+        with tempfile.NamedTemporaryFile('wb', prefix='tmp_mne_prog') as tf:
+            self._mmap_fname = tf.name
+        del tf  # should remove the file
+        self._mmap = None
 
     def update(self, cur_value, mesg=None):
         """Update progressbar with current value of process.
@@ -1760,7 +1781,7 @@ class ProgressBar(object):
         self.cur_time = cur_time
         self.cur_value = cur_value
         self.cur_rate = cur_rate
-        progress = min(float(self.cur_value) / self.max_value, 1.)
+        progress = min(float(self.cur_value) / float(self.max_value), 1.)
         num_chars = int(progress * self.max_chars)
         num_left = self.max_chars - num_chars
 
@@ -1813,6 +1834,65 @@ class ProgressBar(object):
         for obj in self.iterable:
             yield obj
             self.update_with_increment_value(1)
+
+    def __call__(self, seq):
+        """Call the ProgressBar in a joblib-friendly way."""
+        while True:
+            try:
+                yield next(seq)
+            except StopIteration:
+                return
+            else:
+                self.update_with_increment_value(1)
+
+    def subset(self, idx):
+        """Make a joblib-friendly index subset updater.
+
+        Parameters
+        ----------
+        idx : ndarray
+            List of indices for this subset.
+
+        Returns
+        -------
+        updater : instance of PBSubsetUpdater
+            Class with a ``.update(ii)`` method.
+        """
+        return _PBSubsetUpdater(self, idx)
+
+    def __setitem__(self, idx, val):
+        """Use alternative, mmap-based incrementing (max_value must be int)."""
+        if not self._do_print:
+            return
+        assert val is True
+        self._mmap[idx] = True
+        self.update(self._mmap.sum())
+
+    def __enter__(self):  # noqa: D105
+        if op.isfile(self._mmap_fname):
+            os.remove(self._mmap_fname)
+        # prevent corner cases where self.max_value == 0
+        self._mmap = np.memmap(self._mmap_fname, bool, 'w+',
+                               shape=max(self.max_value, 1))
+        return self
+
+    def __exit__(self, type, value, traceback):  # noqa: D105
+        """Clean up memmapped file."""
+        # we can't put this in __del__ b/c then each worker will delete the
+        # file, which is not so good
+        self._mmap = None
+        os.remove(self._mmap_fname)
+        print('')
+
+
+class _PBSubsetUpdater(object):
+
+    def __init__(self, pb, idx):
+        self.pb = pb
+        self.idx = idx
+
+    def update(self, ii):
+        self.pb[self.idx[:ii]] = True
 
 
 def _get_terminal_width():